removed custom command's adding step and make it inline
